test: extract interaction and plan suites
